+++ Core/HotfixMods.Core/Models/TrinityCore/CharacterCustomizations.cs
@@ -5,8 +5,8 @@ namespace HotfixMods.Core.Models.TrinityCore
     [CharactersSchema]
     public class CharacterCustomizations 
     {
-        public int Guid { get; set; }
-        public int ChrCustomizationOptionId { get; set; }
-        public int ChrCustomizationChoiceId { get; set; }
+        public ulong Guid { get; set; }
+        public uint ChrCustomizationOptionId { get; set; }
+        public uint ChrCustomizationChoiceId { get; set; }
     }
 }
