+++ Core/HotfixMods.Core/Models/TrinityCore/Characters.cs
@@ -5,6 +5,6 @@ namespace HotfixMods.Core.Models.TrinityCore
     [CharactersSchema]
     public class Characters 
     {
-        public int Level { get; set; }
+        public byte Level { get; set; }
     }
 }
