+++ Core/HotfixMods.Core/Models/TrinityCore/Creature.cs
@@ -5,7 +5,7 @@ namespace HotfixMods.Core.Models.TrinityCore
     [WorldSchema]
     public class Creature
     {
-        public int Guid { get; set; }
+        public ulong Guid { get; set; }
         public int VerifiedBuild { get; set; }
     }
 }
